Move src/utils/rate-limit to src/core/types/function
